Allow import of Tag groups.
SB has a thing which it calls tag groups.  Users can set these up
and populate them with a number of pupils - only pupils.  It is
useful to be able to attach these to events, so the import code
now imports them.  This involved creating a new type of Group - Taggroup.