2 new flags - commit and notify
Some users are allowed to create, for instance, a calendar entry
without getting approval.  They are not however able to approve other
people's entries.  New flag to allow this.

Also added a flag to control e-mail notifications, but the notifications
themselves aren't quite there yet.

Fixed a problem whereby event durations could not be changed by dragging.
To do this, I had to disable dynamic re-sizing of the calendar.  Need to
work out how to re-enable that later.