Fix display of events to non logged-in users.
A small problem arose from the recent changes for approving commitments.
If an event was in the public calendar, but incomplete in terms of
other resources which it was expecting, then the display of a calendar
view for a non logged-in user failed - the user got no events back at
all.  This release fixes that problem.