Fix identification of hiatuses in import code.
There was a problem with the suspension of 5th year lessons at the
start of next term, in that it was causing a couple of assemblies
to be suspended, even though they involved other pupils apart from
5th years.  This was because the attachement of potential hiatuses
to events was happening before the merging of compound events.  By
switching the order round the problem is corrected.