Improvements to timetable loading
Main changes are that you can now choose which timetable to load,
and it deletes any entries in the d/b which are there but
shouldn't be.

Tidy up an extension of how the command line is parsed.  Cope
with case where iSAMS sends inconsistent information - tutor
groups but no corresponding tutors.