Added --ahead option.
Sometimes it is necessary to do test loads of a year's data
before the supplying MIS has been rolled over.  This then
means that the pupils appear to be in the wrong academic
years.  The MIS thinks they're in the one just ending, but
Scheduler wants to know what year they're going to be in.

By specifying "--ahead 1" you can tweak their academic years
to make things work smoothly.