Delay identifying tutor id for tutor group.
Hit a problem with when the loader was trying to find the
database record for the tutor to attach to a tutor group.
It was doing it at initialisation time, and if the tutor
was a brand new member of staff then it failed because he
hadn't been loaded yet.

Now delay that step until we're actually loading the tutor
groups, by which time staff records have been created.