Recognize subjects in lower school lessons.
There is a design error in the iSAMS API which they refuse to fix.
You can't join up schedule entries for lessons taught by form
(or tutor group in our case) to the forms, or their subjects.
This commit implements a workaround to enable us to spot
the subject associated with each such lesson.