Fix to cope with inconsistent iSAMS d/b.
It has become apparent that the iSAMS database and application
lack integrity checks in the area of pupil membership of teaching
groups.  We started getting data down which specified the same
pupil being a member of a particular teaching group three times.

Our end coped fine because we have proper integrity checks, but it
did produce some warning messages.  Now explicitly check for this
error in the data, and log it unless the --quiet option is specified.