Add property to all covered lessons.
Provided a property called "Covered" exists in the system it will
be added to any event for which cover has been arranged.  Similarly,
it will be removed if the cover is removed again.  A maintenance
method has been added to the Concern model to enable all existing
covered events to be so flagged.