Cope with orphaned location alias.
It is possible for a location alias to exist with no corresponding
location.  Cope explicitly with this case.

(It did sort of work before, but not in the extreme case of there
being a location alias without a location, and a compound element
name which was exactly the same string.)