Dump entire timetable to CSV file.
OTL came up with a request to have the entire timetable in a
spreadsheet so he could do some analysis on it.  Since the import
utility already contained the timetable as an array this was
relatively straightforward to add.

Also removed a spurious "loader" parameter from the Group#yeargroup
methods, and fixed the identification of subjects, which previously
worked only for a subset of subjects.