Reduce notification e-mails.
Due to a slight inconsistency in the code, approvers for resources
would continue to get a daily e-mail for any rejected events in the
system, but not including those rejected events.  Thus, if there
were no real requests pending, the user got an empty e-mail.

Code is now more consistent and won't send an e-mail unless there
are pending requests.

Also fixed the script which runs it so it will dynamically decide
whether it is on a live or development system.