Show pending events to requesters.
In the past, pending requests have been shown in the context of
the requested element only to the approvers and system admins.
This causes confusion to people who enter only the requested
resource and nothing else into their event - they put it in and
immediately they can't see it.  (They have of course un-ticked
the "My events" box.)

With this change, they can see it in the context of the resource
which they've requested, albeit greyed out.