Set existing staff records to active.
The importer takes a very conservative approach to changing the "active"
flag on records.  Historically it has set it on a best-effort basis, and
then refused to change it.  This has led to one slight problem, in that
staff get added in an incomplete fashion to iSAMS (resulting in a record
in Scheduler with active set to false) and then don't automatically
change when the record is completed.

The method now is to allow the flag to be changed from false to true,
but never the other way around.