Fix to display of unapproved events.
Code was added in v0.2 of the software to allow requesters to see
events in the context of a requested resource even before the
request was approved.  Users were getting confused by not being
able to see them.

Unfortunately this led to a display problem in the case of one very
edge condition, resulting in an event not being displayed.  This
commit fixes that problem.