Correct d/b query for fetching notes to display
The rules about whether a not will be displayed are slightly
complex in the case of notes attached to commitments.  They are
shown to the requester, and to the person/people who might
approve the commitment, but not generally to anyone else.

The d/b query to achieve this was slightly wrong.  It used
and INNER JOIN where it should have used a LEFT JOIN.  Corrected.