Add check on cover of inactive staff.
iSAMS lacks the concept of invigilation, and so it has to be done
by setting up dummy lessons with dummy staff and covering them.

The dummy staff weren't active in Scheduler, which caused problems
for the process of importing cover.  Added an extra check to detect
this in the future.