Correct previous fix - use d/b idea of active.
The previous hot fix wasn't quite right.  To check whether a member
of staff was active it was using the loader's calculated value,
where it should have been using the database's actual value.  It
can happen that the former is "false" whilst the latter is "true"
(but not the other way around) when a staff member has been set
to active manually.