Fixed auto-complete.
I've spent ages on this problem - for some reason the auto-complete
code wasn't shoving an ID into the id field.  All down to a missing
hash.  I was giving the id of the field to fill in, where I should
have been giving a CSS selector.  CSS selectors for IDs involve a
leading hash character.