About to split request view.
It's become apparent that I need to split the request view into
two separate views with a controlling overview.  I need to fetch
and display the basic request data quite quickly, and refresh it
as time goes on, but I also need to request a list of candidates
which is relatively slow.  I never want to refresh the list of
candidates (too slow), but I want to get stuff up on screen before
the list of candidates arrives.