Reduce import warnings.
It is becoming increasingly apparent that the iSAMS database lacks
basic integrity checking.  We're getting steadily more instances of,
for instance, pupils being members of groups at times when the
groups in theory do not exist.  Reduce the warning messages which
the importer produces about this.

Also noticed that the script invoking the importer was using a
hard-coded directory path where it could have been using an
existing environment variable.  Fixed that.