Fix non-functional date entry boxes.
This is a long-standing bug.  When the event manipulation dialogue
was split into two separate frames - display and editing - the
date entry boxes stopped working properly for editing.  They still
worked fine when creating a new event, but not for editing.  The
user was left to do all the work.

This commit fixes that problem.