Fix adding of concern which already exists.
It is possible that a user will ask to view something (like the
Calendar) for which they already have a concern (a coloured
box on the LHS) but which is turned off.

In this circumstance what the program should do is not to add
it again, but simply to enable it.  This used to happen but
was broken in about release 0.15.4.  Now works again.