Fix for ical feed specified by initials
When a user requests an ical feed by initials, and the school
has used the same initials for a previous member of staff, the
system was using the first of the two - i.e. the old member
of staff.  It now selects only current members of staff, so
provided you don't have two simultaneous members of staff with
the same initials then you're fine.