Tidy functions and data in schedule.js
After recent tidying changes to the CoffeeScript which generates
the JavaScript for the main Scheduler page, I realised that I'd
ended up with some pointless stubs of functions.  They existed
merely to call more functions with exactly the same parameters.
Removed them, and reduced the use of accidentally global variables.