Reduce invigilation notice headers.
Staff were being slightly confused by the invigilation notifications
which included the date and time of the notice in the header.
As it was the first thing they saw, they assumed it was the time of
the invigilation.  It's unnecessary anyway, because the e-mail
automatically carries the time of sending as part of the standard
content of an e-mail.  Taken out.