Efficiency improved.
By pre-loading elements and entitities, we dispose of the n+1.

There are still unnecessary queries being done, because we pre-load
locations and properties too even though they aren't needed.  OTOH,
these queries take OTOO 0.1ms, and we have already reduced the
total query time for a large week's worth of events (all the
invigilation in a week) from 500ms to 200ms with this commit.