Correct switch events to/from timed.
When an event is timed and and becomes all-day, or vice versa,
there used to be some quite complex (and broken) processing
involving setting times to 08:00 by default.

Now we display multi-day timed events rather better, this is
unnecessary and has been simplified.

Also fixed an error in the display of multi-day timed events which
end at exactly midnight.