Add warning message for no resources.
The most common cause for confusion for new users is if they create
an event but add no resources to it.  They're then puzzled as to
why no-one can see it.  The documentation emphasises this as a
potential point of confusion, but it seems that people just don't
read documentation.

This commit adds a warning message which appears when the user is
editing an event which has no resources.  As soon as they add a resource
the message disappears.