Show warning in event display too.
The warning originally appeared only in the edit dialogue, but
now it appears in the show dialogue too, provided the user has
the potential ability to do something about it - i.e. the means
to edit the event.

Also renamed the control flag in the user record to remove
ambiguity.

Not sure whether this new place to display the message should not
completely replace the old one.  Or perhaps make them system
configurable?