Display warning only on "Show" screen.
The resource warning message originally appeared on the event
editing screen.  Then I added it to the show screen.  Then I
decided that that was a better place for it overall, so it's
now disabled on the edit screen.

A user who creates an event in the expected way will thus
not see the message at all.  Only a user who exits the edit
without adding any resources will see the message.