Re-arrange JavaScript assets to circumvent bug
There is a bug in formBuilder, in that it injects a hard-coded
CSS stylesheet into your application without permission, and
without even being invoked.  Just including the JS file is
enough to do the damage.

Have re-arranged the assets in the application so that FB is
included only where it absolutely has to be.