Allow array of pre-committed elements on event.
We have for a long time allowed a pre-committed element when
creating an event.  The user drags an element onto the display.
Its ID is noted and embedded in a hidden field in the event
creation form.  When the form comes back, the value is retrieved
and the element is added to the event.

This commit enhances the controller to allow an array of comma-separated
element IDs to be passed back in that field.  It is thus possible for
JavaScript in the front end to pack more in there as a result of the
user ticking tick-boxes on pre-requisites.