Try to fix e-mail formatting issues.
When deployed to live, certain problems became apparent with e-mails
which are viewed through Gmail.  Although you can show HTML
e-mails there, it does quite a lot of messing about with both
your HTML and your CSS, resulting in them not looking quite the
same as when they are viewed in their virgin state.

This commit addresses two points:

* Try to make the Scheduler link in the header appear in white and
  without adornment.

* Add a "clear" property, to stop the buttons cascading inwards in
  the resource approvals listing.