Improve capitalisation of staff surnames.
The only definitive source of staff surnames unfortunately
has them in all caps.  They do exist elsewhere, but not
in a reliable place.  The code now takes the all-caps
version and looks for the same string elsewhere, but not
necessarily in the same case.  If found, that is used
instead.