Handle repeating event in holiday week sensibly
If the user creates an event in a holiday week and then sets it
to repeat, it was by default deleted because the default setting
was term weeks only.

Now if the initial event is in term time, the default is just term
time, but if the initial event is outside term, then the default
repetition is all weeks - even holiday weeks.