Don't set group to non-current if ending in the future
The method Group#ceases_existence() by default assumes you want
it to cease existence today and thus sets the date and sets the
current flag to false.

It is however possible to call it with an explicit date, and if
that date is in the future then we don't want the current flag
immediately to go to false.  It will be set to false by the
nightly cron job at the appropriate time.