Fix double creation of elements
A recent change had the knock-on effect that creating a new
location alias without a linked location caused a new location
to be created (correct behaviour) but with two element records.

This is because the update_element method on the entity was
being called twice and didn't cope with it.  Fixed the update_element
method so it does cope with being called twice.