Fix Cranford-specific lesson merging.
Two issues.  When consecutive lessons are being merged, the algorithm
for doing it was modifying an array whilst iterating through the same
array.  Ruby gets confused if you do this.

Now saves all the modifications to the array for after the iteration.

Also added a custom hash_key to the PASS_ScheduleEntry model to allow
more lessons to be merged at Cranford.