Give link to event on search result page.
Also change the logic slightly, so that non logged-in users
see events with firm commitments to public categories, rather
than events which are complete and linked to public categories.

When there was only the public calendar to worry about, the
two were equivalent, but now an event might be confirmed in the
public calendar, but still lacking other resources.