Don't delete and re-create attachments when note edited
Previously, each time a note's body was edited, we deleted
all the notes file attachments and re-created them.  Since
we now have additional processing when an attachment is
deleted this was getting very messy.

The relevant function is now more intelligent and deletes
only the attachments it really wants to be rid of.