In event deleted email, don't repeat resource
When an event is deleted, the subject line tells you about
what resource the message relates to.  This was being repeated
in the body, which gave formatting issues and meant we
included the same information twice.

Now in the body it just says "Event deleted", but if there
is a relevant resource then it's still in the subject line
(and indeed, further down within the body).