Fix erroneous SQL injection warning.
Slightly odd feature has been added in current version of Rails.

It issues a warning if you use LENGTH("FIXED STRING") in a SQL
query, saying there is a risk of SQL injection.  This is silly - it's
a hard-coded string.  I've had to add "Arel.sql()" around my strings
to get it to stop.