Fix bug in free staff output for combined weeks
A logic error in the utility for producing a listing of free
staff at specified times led to some nonsensical output.

In the code for handling both weeks together, if very few staff
(< 10) were busy then the code tried to list the few who were
unavailable but instead listed everyone.  This commit fixes
that bug.