Fix bug in deciding whether events overlap.
Event start times are stored in the database including zone information,
so an all-day event in the UK in the summer is stored as starting
at 23:00 GMT of the day before.  If you then pass in a pure date
to a SQL selector you will get a false overlap.  Need to make sure
that the SQL generated uses the same convention.