Cope with references to deleted pupil courses
A bug in the server meant that it sent serialized JSON which still
contained references to pupil courses which had been deleted.  This
led to a failure in the JavaScript at initialization.  That bug
at the server end has been fixed, but this commit makes the JavaScript
now cope even if the bug is there.