Add more info for erroneous cover records
iSAMS seems to lack some basic integrity checking which results
in us getting gash information.  One particular problem is that
cover records for a lesson do not get removed when the parent
lesson is removed.  We cope with this and simply log the discrepancy.

Here we're just adding a bit more info about the error.