Fix error in setting of element_owner flag
Code in the user model tried to keep the element_owner flag updated
in the user record as concerns changed but due to a stupid typo it
wasn't actually doing its job.

The test for this functionality failed to detect the problem because
it was looking at the in memory copy of the user record and not
going back to the database.  Fixed that too.